+++ src/query_search/blog_search.clj
@@ -1,19 +1,17 @@
 (ns query-search.blog-search
   "Модуль поиска по блогам."
-  (:require [query-search.crawler :as crawler]
-            [query-search.settings :as settings]
-            [query-search.logger :refer :all]))
+  (:require [clojure.algo.monads :refer [reader-m asks domonad]]
+            [query-search.crawler :as crawler]
+            [query-search.logger :refer [log]]))
 
 (defn search
   "Поиск в блогах по ключевым словам."
   [keywords]
-  (log "Ищем в блогах по ключевым словам:" keywords)
-  (crawler/crawl
-    (map
-      #(spy
-         "Сформировали запрос для поиска по блогам:"
-         (hash-map
-           :url (settings/get-setting :api-url)
+  (domonad reader-m [settings (asks :settings)]
+    (crawler/crawl
+      (map
+        #(hash-map
+           :url (:api-url settings)
            :params {"text" %
-                    "numdoc" (settings/get-setting :numdoc)}))
-      keywords)))
+                    "numdoc" (:numdoc settings)})
+        keywords))))