+++ src/query_search/parser.clj
@@ -6,15 +6,14 @@
 (defn- get-links
   "Достаёт ссылки из ответа."
   [xml]
-  (map #(:text %) (xpath/$x "/rss/channel/item/link" xml)))
+  (map :text (xpath/$x "/rss/channel/item/link" xml)))
 
 (defn- get-domains
   "Достаёт домены из ссылок."
   [links]
-  (map #(vec [(re-find #"\w+\.\w+(?=[/?#])" (str % "/")) %]) links))
+  (map #(vector (re-find #"\w+\.\w+(?=[/?#])" (str % "/")) %) links))
 
 (defn parse
   "Извлекает домены из XML."
   [xml]
-  (log "Извлекаем домены из XML")
   (get-domains (get-links xml)))