+++ test/query_search/blog_search_test.clj
@@ -2,22 +2,16 @@
   "Тестирование поиска по блогам."
   (:require [clojure.test :refer :all]
             [clojure.string :refer [includes?]]
-            [query-search.testing.fake-server :as server]
-            [query-search.blog-search :refer [search]]))
-
-
-(defn fake-server-fixture
-  "Fixture для тестирования API.
-   Запускает и останавливает API-сервер."
-  [f]
-  (let [stop (server/start)]
-    (f)
-    (stop :timeout 500)))
+            [query-search.testing.fixtures :refer [fake-server-fixture]]
+            [query-search.blog-search :refer [search]]
+            [query-search.settings :refer [settings]]))
 
 
 (deftest search-test
   (testing "Поиск по блогам."
-    (is (includes? (first @(search ["scala"])) "yablogs:author"))))
+    (let [testing-server-url (str "http://localhost:" (:testing-server-port settings) "/")
+          env {:settings (assoc settings :api-url testing-server-url)}]
+      (is (includes? (first @((search ["scala"]) env)) "yablogs:author")))))
 
 
 ;;; Используем фейковый API-сервер для тестирования